feat: create standalone compact card view with view model
- Moves CompactCardView from EventView.swift into its own dedicated file to improve modularity.
- Introduces CompactCardViewModel to encapsulate presentation logic, such as formatting dates and prices.
- Refactors CompactCardView to consume data from the new view model, separating view concerns from data logic.
