+++ Event Tracker/Event Tracker/Features/Events/Views/EventView.swift
@@ -237,72 +237,6 @@ struct EmptyEventsView: View {
     }
 }
 
-struct CompactCardView: View {
-    let event: CreateEventModel
-    @EnvironmentObject private var router: Router
-    
-    var body: some View {
-        HStack(spacing: 12) {
-            AsyncImage(url: URL(string: event.images.first?.thumbnailUrl ?? "")) { image in
-                image
-                    .resizable()
-                    .aspectRatio(contentMode: .fill)
-            } placeholder: {
-                RoundedRectangle(cornerRadius: 8)
-                    .fill(Color(.systemGray5))
-                    .overlay(
-                        Image(systemName: "photo")
-                            .foregroundColor(.gray)
-                    )
-            }
-            .frame(width: 60, height: 60)
-            .clipShape(RoundedRectangle(cornerRadius: 8))
-            
-            VStack(alignment: .leading, spacing: 4) {
-                Text(event.title)
-                    .font(.subheadline)
-                    .fontWeight(.semibold)
-                    .lineLimit(1)
-                
-                HStack {
-                    
-                    Text(event.startDate, style: .date)
-                        .font(.caption)
-                        .foregroundColor(.secondary)
-                    
-                    Spacer()
-                    
-                    Text(event.pricing.formattedPrice)
-                        .font(.caption)
-                        .fontWeight(.medium)
-                        .foregroundColor(event.pricing.isFree ? .green : .blue)
-                        .padding(.trailing, 10)
-                    
-//                    if !event.participants.isFull {
-//                        Text("\(event.participants.remainingSpots) yer kaldı")
-//                            .font(.caption)
-//                            .foregroundColor(.orange)
-//                    } else {
-//                        Text("Dolu")
-//                            .font(.caption)
-//                            .foregroundColor(.red)
-//                    }
-                }
-                
-            }
-            
-        }
-        .padding(.horizontal, 5)
-        .padding(.vertical, 5)
-        .background(Color(.systemGray6))
-        .cornerRadius(12)
-        .onTapGesture {
-            router.push(.eventDetail(event))
-        }
-    }
-}
-
-
 #Preview {
     EventView()
 }
